Use more aggressive scheduling in default implementation
As we are casting next event period to integer milliseconds, there's a
50% chance we will round down and wake up too early, and have to
schedule another tick in 1ms.

However, we may in fact achieve sub-ms scheduling, as event callback
calls self to catch up on any events that might have happened, before
resetting the timer.

All in all, I think this is a worthwhile change.